+++ src/main/java/com/example/prography_project/user/dto/UserListResponseDto.java
@@ -8,12 +8,12 @@ import java.util.List;
 
 @Getter
 public class UserListResponseDto {
-    private long totalElements;
+    private int totalElements;
     private int totalPages;
     private List<UserResponseDto> userList;
 
     public UserListResponseDto(Page<UserResponseDto> userPage) {
-        this.totalElements = userPage.getTotalElements();
+        this.totalElements = (int) userPage.getTotalElements();
         this.totalPages = userPage.getTotalPages();
         this.userList = userPage.getContent();
     }
